Изменил getDefault в Managers; перенес статрый пользовательский сценарий из Main в файл old_main.txt; метод main из FileBackedTaskManager перенес в Main
