+++ old_main.txt
@@ -1,0 +1,58 @@
+public static void main(String[] args) {
+        // Дополнительное задание. Реализуем пользовательский сценарий
+
+        // 1. Создайте две задачи, эпик с тремя подзадачами и эпик без подзадач.
+        TaskManager taskManager = Managers.getDefault();
+        Task task1 = new Task("Task1", "");
+        Task task2 = new Task("Task2", "");
+        taskManager.addTask(task1);
+        taskManager.addTask(task2);
+        EpicTask epicTask1 = new EpicTask("EpicTask1", "");
+        taskManager.addEpicTask(epicTask1);
+        SubTask subTask1 = new SubTask("subTask1", "from epicTask1", epicTask1.getId());
+        SubTask subTask2 = new SubTask("subTask2", "from epicTask1", epicTask1.getId());
+        SubTask subTask3 = new SubTask("subTask3", "from epicTask1", epicTask1.getId());
+        taskManager.addSubTask(subTask1);
+        taskManager.addSubTask(subTask2);
+        taskManager.addSubTask(subTask3);
+        EpicTask epicTask2 = new EpicTask("EpicTask2", "");
+        taskManager.addEpicTask(epicTask2);
+
+        //2. Запросите созданные задачи несколько раз в разном порядке.
+        //3. После каждого запроса выведите историю и убедитесь, что в ней нет повторов.
+        taskManager.getTask(task2.getId());
+        System.out.println("В истории должно быть - Task2\n" + taskManager.getHistory());
+        taskManager.getTask(task2.getId());
+        System.out.println("В истории должно быть - Task2\n" + taskManager.getHistory());
+        taskManager.getEpicTask(epicTask2.getId());
+        System.out.println("В истории должно быть - Task2 -> epicTask2\n" + taskManager.getHistory());
+        taskManager.getEpicTask(epicTask2.getId());
+        System.out.println("В истории должно быть - Task2 -> epicTask2\n" + taskManager.getHistory());
+        taskManager.getTask(task2.getId());
+        System.out.println("В истории должно быть - epicTask2 -> Task2\n" + taskManager.getHistory());
+        taskManager.getSubTask(subTask3.getId());
+        System.out.println("В истории должно быть - epicTask2 -> Task2 -> subTask3\n" + taskManager.getHistory());
+        taskManager.getEpicTask(epicTask2.getId());
+        System.out.println("В истории должно быть - Task2 -> SubTask3 -> epicTask2\n" + taskManager.getHistory());
+        taskManager.getSubTask(subTask3.getId());
+        System.out.println("В истории должно быть - Task2 -> epicTask2 -> SubTask3\n" + taskManager.getHistory());
+
+        //4. Удалите задачу, которая есть в истории, и проверьте, что при печати она не будет выводиться.
+        taskManager.removeSubTask(subTask3.getId());
+        System.out.println("В истории должно быть - Task2 -> epicTask2\n" + taskManager.getHistory());
+        taskManager.removeEpicTask(epicTask2.getId());
+        System.out.println("В истории должно быть - Task2\n" + taskManager.getHistory());
+        taskManager.removeTask(task2.getId());
+        System.out.println("В истории должно быть пусто\n" + taskManager.getHistory());
+
+        //5 Удалите эпик с тремя подзадачами и убедитесь, что из истории удалился как сам эпик, так и все его подзадачи.
+        taskManager.getEpicTask(epicTask1.getId());
+        System.out.println("В истории должно быть - epicTask1\n" + taskManager.getHistory());
+        taskManager.addSubTask(subTask3); // ранее удалили, добавляем заново
+        taskManager.getSubTask(subTask1.getId());
+        taskManager.getSubTask(subTask2.getId());
+        taskManager.getSubTask(subTask3.getId());
+        System.out.println("В истории должно быть - epicTask1 и все его подзадачи\n" + taskManager.getHistory());
+        taskManager.removeEpicTask(epicTask1.getId());
+        System.out.println("В истории должно быть пусто\n" + taskManager.getHistory());
+    }
