+++ test/util/ManagersTest.java
@@ -4,13 +4,18 @@ import org.junit.jupiter.api.Test;
 import service.interfaces.HistoryManager;
 import service.interfaces.TaskManager;
 
+import java.io.IOException;
+import java.nio.file.Files;
+import java.nio.file.Path;
+
 import static org.junit.jupiter.api.Assertions.assertNotNull;
 
 class ManagersTest {
 
     @Test
-    void getDefault() {
-        TaskManager aDefault = Managers.getDefault();
+    void getDefault() throws IOException {
+        Path tempFile = Files.createTempFile("test", ".txt");
+        TaskManager aDefault = Managers.getDefault(tempFile);
         assertNotNull(aDefault);
     }
